move landing into seed
