+++ seed/landing/features/steps.py
@@ -4,7 +4,7 @@
 from salad.steps.everything import *
 from lettuce import step
 from django.core.urlresolvers import reverse
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from superperms.orgs.models import Organization
