+++ seed/landing/management/commands/update_eula.py
