+++ seed/landing/static/landing/images/logo-chrome-64.png
