+++ seed/landing/static/landing/images/logo-safari-64.png
