+++ seed/landing/templates/landing/thank_you.html
