+++ seed/landing/templates/tos/tos_check.html
