+++ seed/tasks.py
@@ -23,7 +23,7 @@ from django.core.urlresolvers import reverse_lazy
 from celery.task import task, chord
 from celery.utils.log import get_task_logger
 from seed.audit_logs.models import AuditLog
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from seed.lib.mcm import cleaners, mapper, reader
 from mcm.data.ESPM import espm as espm_schema
 from mcm.data.SEED import seed as seed_schema
