+++ seed/tests/test_account_views.py
@@ -6,7 +6,7 @@ import json
 from django.core.urlresolvers import reverse_lazy
 from django.test import TestCase
 from superperms.orgs.models import Organization
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from seed.views.main import _get_default_org
 from seed.views.accounts import _dict_org, _get_js_role, _get_role_from_js
 from superperms.orgs.models import (
