+++ seed/tests/test_dataset.py
@@ -7,7 +7,7 @@ from django.core.urlresolvers import reverse_lazy
 from django.test import TestCase
 from superperms.orgs.models import Organization, OrganizationUser
 from seed.data_importer.models import ImportFile, ImportRecord
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 
 
 class DeleteFileViewTests(TestCase):
