+++ seed/tests/test_delete.py
@@ -13,7 +13,7 @@ from superperms.orgs.models import (
     Organization,
 )
 from seed.audit_logs.models import AuditLog
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from seed.data_importer.models import ImportRecord
 from seed.tests.util import FakeRequest
 from seed.factory import SEEDFactory
