+++ seed/tests/test_models.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from django.test import TestCase
 from superperms.orgs.models import Organization, OrganizationUser
 from seed.data_importer.models import ImportFile, ImportRecord
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from seed import models as seed_models
 from seed.mappings import mapper
 from seed.tests import util
