+++ seed/tests/test_tasks.py
@@ -11,7 +11,7 @@ from django.test import TestCase
 from django.core.files import File
 from seed.audit_logs.models import AuditLog
 from seed.data_importer.models import ImportFile, ImportRecord
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from superperms.orgs.models import Organization, OrganizationUser
 from seed.models import (
     ASSESSED_RAW,
