+++ seed/tests/tests.py
@@ -2,10 +2,10 @@
 :copyright: (c) 2014 Building Energy Inc
 """
 
-from landing.models import SEEDUser as User
+from seed.landing.models import SEEDUser as User
 from django.test import TestCase
 
-from ..models import Project, Compliance, BuildingSnapshot, CanonicalBuilding
+from seed.models import Project, Compliance, BuildingSnapshot, CanonicalBuilding
 from seed.utils.organizations import create_organization
 from seed.utils.buildings import get_buildings_for_user_count
